Moved BypassPullRequestAllowances type to new location Generated API report
