+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/github/githubRepoPush.ts
@@ -25,7 +25,7 @@ import { parseRepoUrl } from '../publish/util';
 import { getOctokitOptions, initRepoPushAndProtect } from './helpers';
 import * as inputProps from './inputProperties';
 import * as outputProps from './outputProperties';
-import { BypassPullRequestAllowances } from '../../types';
+import { BypassPullRequestAllowances } from './types';
 
 /**
  * Creates a new action that initializes a git repository of the content in the workspace
