+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/publish/github.ts
@@ -29,7 +29,7 @@ import {
 import * as inputProps from '../github/inputProperties';
 import * as outputProps from '../github/outputProperties';
 import { parseRepoUrl } from './util';
-import { BypassPullRequestAllowances } from '../../types';
+import { BypassPullRequestAllowances } from '../github/types';
 /**
  * Creates a new action that initializes a git repository of the content in the workspace
  * and publishes it to GitHub.
